3270 - Print - Cache pdf (#3649)
* 3270 - Update yarn lock

* 3270 - Add country iso to Repository getOne props

* 3270 - Add country_summary getLastUpdate

* 3270 - Add FileRepository removeOne

* 3270 - Omit id and uuid unnecessary properties when creating a repository item

* 3270 - Add pdfRepository controller

* 3270 - Update print report endpoint logic

* 3270 - Fix Report naming and export

* 3270 - Fix yarn.lock

* 3270 - Remove country_iso from Repository.getOne props

* 3270 - Add CountrySummary getOneOrNone

* 3270 - Account for reports per language and with only tables

* 3270 - update force req param

* 3270 - remove FileRepository.removeOne

* 3270 - update file cleanup: call repo.remove only if uuids.length

* 3270 - restore file cleanup repeat pattern

---------

+++ src/server/schedulers/initFileCleanup.ts
@@ -30,11 +30,13 @@ export const initFileCleanup = (connection: IORedis): Worker => {
         `
       const uuids = await client.map(query, [], (row) => row.uuid)
 
-      const files = await FileRepository.removeMany({ uuids })
+      if (uuids.length > 0) {
+        const files = await FileRepository.removeMany({ uuids })
 
-      files.forEach((file) => {
-        Logger.info(`[${name}] removed file ${file.name} (${file.uuid})`)
-      })
+        files.forEach((file) => {
+          Logger.info(`[${name}] removed file ${file.name} (${file.uuid})`)
+        })
+      }
 
       Logger.info(`[${name}] ** terminated`)
     },
